+++ CMakeLists.txt
@@ -58,8 +58,8 @@ option(ENABLE_SRT "Enable SRT" ON)
 option(ENABLE_TESTS "Enable Tests" ON)
 option(ENABLE_SCTP "Enable SCTP" ON)
 option(ENABLE_WEBRTC "Enable WebRTC" ON)
-option(ENABLE_QUIC "Enable QUIC/HTTP3 support" OFF)
-option(ENABLE_LSQUIC_AUTO_BUILD "Automatically download and build a private BoringSSL+lsquic dependency bundle for QUIC" OFF)
+option(ENABLE_QUIC "Enable QUIC/HTTP3 support" ON)
+option(ENABLE_LSQUIC_AUTO_BUILD "Automatically download and build a private BoringSSL+lsquic dependency bundle for QUIC" ON)
 set(LSQUIC_VERSION "v4.6.0" CACHE STRING "LSQUIC git tag/branch used for auto-build")
 set(BORINGSSL_REVISION "9fc1c33e9c21439ce5f87855a6591a9324e569fd" CACHE STRING "BoringSSL git revision used for local auto-build when QUIC is enabled")
 set(GO_AUTO_VERSION "1.24.5" CACHE STRING "Go version used for local auto-build fallback when QUIC dependencies are enabled")
