+++ quic/QuicPluginBackend.h
@@ -8,6 +8,7 @@
 
 #include "Common/MediaSource.h"
 #include "Common/Parser.h"
+#include "Http/HttpBody.h"
 #include "QuicBackend.h"
 #include "QuicPlugin.h"
 
@@ -57,14 +58,32 @@ private:
             return seed;
         }
     };
+    struct RequestKey {
+        uint64_t conn_id = 0;
+        uint64_t stream_id = 0;
+
+        bool operator==(const RequestKey &that) const {
+            return conn_id == that.conn_id && stream_id == that.stream_id;
+        }
+    };
+
+    struct RequestKeyHash {
+        size_t operator()(const RequestKey &key) const {
+            auto seed = std::hash<uint64_t>()(key.conn_id);
+            seed ^= std::hash<uint64_t>()(key.stream_id) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
+            return seed;
+        }
+    };
 
     static RouteKey makeRouteKey(QuicSlice local_ip, uint16_t local_port, QuicSlice peer_ip, uint16_t peer_port);
     static RouteKey makeRouteKey(const QuicPacket &packet);
-    static uint64_t makeRequestKey(uint64_t conn_id, uint64_t stream_id);
+    static RequestKey makeRequestKey(uint64_t conn_id, uint64_t stream_id);
     static bool extractPacketConnectionIds(const uint8_t *data, size_t len,
                                            std::string *dcid, std::string *scid);
     void rememberPacketConnectionIds(const uint8_t *data, size_t len, const toolkit::EventPoller::Ptr &poller);
     void rememberRoute(const QuicPacket &packet);
+    std::shared_ptr<PendingRequest> removeRequest(uint64_t conn_id, uint64_t stream_id);
+    std::vector<std::shared_ptr<PendingRequest>> removeConnectionRequests(uint64_t conn_id);
     void dispatchRequest(const std::shared_ptr<PendingRequest> &request);
     bool checkLiveStream(const std::shared_ptr<PendingRequest> &request, Parser &parser,
                          const std::string &schema, const std::string &url_suffix,
@@ -72,6 +91,11 @@ private:
     bool checkLiveStreamFlv(const std::shared_ptr<PendingRequest> &request, Parser &parser);
     bool checkLiveStreamTS(const std::shared_ptr<PendingRequest> &request, Parser &parser);
     bool checkLiveStreamFMP4(const std::shared_ptr<PendingRequest> &request, Parser &parser);
+    static toolkit::Any makeSenderInfo(const std::weak_ptr<PendingRequest> &weak_request);
+    void sendHttpBody(const std::shared_ptr<PendingRequest> &request,
+                      const HttpBody::Ptr &body, size_t chunk_size);
+    void sendTextResponse(const std::shared_ptr<PendingRequest> &request, int code,
+                          const std::string &content_type, const std::string &text);
     void sendResponseHeaders(const std::shared_ptr<PendingRequest> &request, int code,
                              const StrCaseMap &header_out, int64_t body_size,
                              bool no_content_length, bool fin);
@@ -86,7 +110,7 @@ private:
     mutable std::mutex _route_mutex;
     std::unordered_map<RouteKey, std::weak_ptr<toolkit::Socket>, RouteKeyHash> _route_map;
     mutable std::mutex _request_mutex;
-    std::unordered_map<uint64_t, std::shared_ptr<PendingRequest>> _request_map;
+    std::unordered_map<RequestKey, std::shared_ptr<PendingRequest>, RequestKeyHash> _request_map;
 };
 
 } // namespace mediakit
