+++ quic/plugin/lsquic/LsquicClientEngine.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../../QuicPluginABI.h"
+
+namespace mediakit {
+
+IQuicClientEngine *createLsquicClientEngine(IQuicClientHost &host, const QuicClientConfig &config);
+
+} // namespace mediakit
