+++ quic/plugin/lsquic/LsquicServerEngine.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "../../QuicPluginABI.h"
+
+namespace mediakit {
+
+IQuicServerEngine *createLsquicServerEngine(IQuicServerHost &host, const QuicServerConfig &config);
+
+} // namespace mediakit
